Skip App service in enforcePublicAuthorization
PR #2891 removed the App.Invoke/InvokeGraphQL skip from
enforcePublicAuthorization, so all App service methods now go through
runAuthorizationCheck with Type="app". The authz model has no "app"
resource type (it uses per-app types like aiSpendTracker), so every
app call returns "access denied".

Skip the entire App service in enforcePublicAuthorization. The broker
handles per-operation CheckAccess for all App methods with correct
resource-type resolution.
